+++ locals.tf
@@ -85,6 +85,6 @@ locals {
   private_dns_label   = "private"
   default_private_dns = cidrhost(cidrsubnet(var.vcn_cidr, var.private_newbits, var.private_netnum), var.lab_dns_hostnum)
   tvd_private_dns     = var.tvd_private_dns == "default" ? local.default_private_dns : var.tvd_private_dns
-  custom_dns_servers  = length(var.custom_dns_servers) == 0 ? [local.tvd_private_dns, var.tvd_public_dns] : var.custom_dns_servers
+  custom_dns_servers  = length(var.custom_dns_servers) == 0 ? [local.tvd_private_dns, var.lab_public_dns] : var.custom_dns_servers
 }
 # --- EOF -------------------------------------------------------------------