+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename *tvd_participants* to *numberOf_labs*
 - rename *tvd_domain* to *lab_domain*
 - rename *tvd_dns_hostnum* to *lab_dns_hostnum*
+- rename *tvd_public_dns* to *lab_public_dns*
 
 ## [3.3.0] - 2023-03-10